+++ drivers/sqlboiler-sqlite3/driver/sqlite3_test.go
@@ -1,0 +1,91 @@
+package driver
+
+import (
+	"bytes"
+	"encoding/json"
+	"flag"
+	"fmt"
+	"io/ioutil"
+	"math/rand"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+	"github.com/volatiletech/sqlboiler/v4/drivers"
+	_ "modernc.org/sqlite"
+)
+
+var (
+	flagOverwriteGolden = flag.Bool("overwrite-golden", false, "Overwrite the golden file with the current execution results")
+)
+
+func TestDriver(t *testing.T) {
+	rand.Seed(time.Now().Unix())
+	b, err := ioutil.ReadFile("testdatabase.sql")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	tmpName := filepath.Join(os.TempDir(), fmt.Sprintf("sqlboiler-sqlite3-%d.sql", rand.Int()))
+
+	out := &bytes.Buffer{}
+	createDB := exec.Command("sqlite3", tmpName)
+	createDB.Stdout = out
+	createDB.Stderr = out
+	createDB.Stdin = bytes.NewReader(b)
+
+	t.Log("sqlite file:", tmpName)
+	if err := createDB.Run(); err != nil {
+		t.Logf("sqlite output:\n%s\n", out.Bytes())
+		t.Fatal(err)
+	}
+	t.Logf("sqlite output:\n%s\n", out.Bytes())
+
+	tests := []struct {
+		name       string
+		config     drivers.Config
+		goldenJson string
+	}{
+		{
+			name: "default",
+			config: drivers.Config{
+				"dbname": tmpName,
+			},
+			goldenJson: "sqlite3.golden.json",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			s := &SQLiteDriver{}
+			info, err := s.Assemble(tt.config)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			got, err := json.MarshalIndent(info, "", "\t")
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if *flagOverwriteGolden {
+				if err = ioutil.WriteFile(tt.goldenJson, got, 0664); err != nil {
+					t.Fatal(err)
+				}
+				t.Log("wrote:", string(got))
+				return
+			}
+
+			want, err := ioutil.ReadFile(tt.goldenJson)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			require.JSONEq(t, string(want), string(got))
+		})
+	}
+
+}